Decompress gziped body before logging (#13)
The request body was logged directly, as raw data, this is problematic if the body is compressed or is not just ASCII data.

This commit fixes it by reading the Content-Encoding header and decompressing if it is gzip.

Logging is disabled by default and a flag (verbose) to enable it is added.